feat: random char on connect

+++ crates/lorenzo/src/image.rs
@@ -3,7 +3,7 @@ pub struct Image {
     pub id: u32,
 
     /// file number corresponding to the image
-    pub file: u64,
+    pub file: u32,
 
     /// source rect, sometimes the image is a portion of a bigger one
     pub x: u16,
